# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jatek Food Delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/objects/{path}:
    get:
      operationId: getObject
      tags:
        - storage
      summary: Serve an uploaded object
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object bytes
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to serve object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /storage/uploads/image:
    post:
      operationId: uploadImage
      tags:
        - storage
      summary: Upload and validate an image
      description: "RBAC: admin, super_admin, manager, or restaurant_owner. The server validates image bytes before storage."
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
          image/png:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
          image/gif:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Image uploaded and validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadResponse"
        "400":
          description: Invalid image bytes or declared MIME type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Image exceeds the 5 MB limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Unsupported image type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to upload image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - bearerAuth: []
...
  /users:
    get:
      operationId: listUsers
      tags:
        - users
      summary: List users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags:
        - users
      summary: Get a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags:
        - users
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      security:
        - bearerAuth: []
      description: JWT required. Users can update themselves; admins can update any user.
    delete:
      operationId: deleteUser
      tags:
        - users
      summary: Delete a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
      security:
        - bearerAuth: []
      description: JWT required. Users can delete themselves; admins can delete any user.
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type:
            - string
            - "null"
        address:
          type:
            - string
            - "null"
        avatarUrl:
          type:
            - string
            - "null"
        isActive:
          type: boolean
        loyaltyPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
        assignedShopId:
          type:
            - integer
            - "null"
      required:
        - id
        - name
        - email
        - role
        - isActive
        - loyaltyPoints
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        avatarUrl:
          type: string
        isActive:
          type: boolean
...
    ImageUploadResponse:
      type: object
      required:
        - url
        - contentType
      properties:
        url:
          type: string
          description: Relative URL that serves the validated image.
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp, image/gif]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by /auth/login, /auth/verify-otp, or /backend/login. Some HTML document endpoints also accept the same token as a token query parameter.